Add twig function to retrieve link type prefix
